Moved night mode out of MainActivity and into ConfigView.
