+++ app/src/main/java/ar/edu/itba/pam/travelapp/main/config/ConfigView.java
@@ -4,4 +4,6 @@ public interface ConfigView {
     void bind();
 
     void handleNightModeSwitch(int mode);
+
+    boolean wasNightModeToggled();
 }
